fix(score): cold-start velocity clamp + TBTS float; topics: niche-tightened prompt (#12)
Three coherent fixes to the scoring + topic-extraction layer.

1. **Cold-start velocity over-suppression** (`pipeline/score.py`)
   Before: a brand-new topic with `mentions_7d=1, mentions_30d=0` returned
   `velocity=0.43`, making fresh signal look dead because `MENTIONS_30D_FLOOR=10`
   was applied uniformly. After: `cold_floor = min(MENTIONS_30D_FLOOR,
   max(mentions_7d, 1))` clamps the floor to the actual 7d count for cold-start
   topics, so the same fresh topic now returns `velocity=1.0`. Established
   topics (mentions_7d ≥ 10) keep the original spike-suppression behavior.

2. **TBTS overplot fix** (`pipeline/score.py` + `pipeline/models.py`)
   Before: `tbts()` returned `round(raw * 100)` as int. ~80% of whisper-stage
   trends collapsed to identical scores, causing Sky Map bubble overplot at
   the same coordinates. After: returns float; `Trend.tbts` Pydantic field
   bumped `int → float`. Near-tied trends now differentiate to one decimal
   place.

3. **Topic-extractor niche bias** (`pipeline/topics.py`)
   Strengthened Claude's system prompt to bias toward multi-source topics
   (arXiv + HN or GitHub + HN) — the topics escaping academic confinement —
   over pure arXiv preprints. Explicitly targets the project's niche lock:
   "AI tools for solo content creators." arXiv-only topics now require 3+
   attributed papers AND creator-explainability before they're kept.

Tests: `tests/test_score.py` adds 2 new tests for the cold-start clamp
(`test_velocity_cold_start_single_mention`, `test_velocity_established_topic_keeps_original_floor`)
and updates 4 existing tests for the float-TBTS return type. Full suite:
467 passed, 2 xfailed (pre-existing), 0 regressions.

+++ pipeline/models.py
@@ -132,7 +132,7 @@ class Trend(BaseModel):
     hidden_gem_score: float
     builder_signal: float
     lifecycle_stage: LifecycleStage
-    tbts: int
+    tbts: float
     convergence: ConvergenceEvent
     summary: str
     summary_confidence: Literal["high", "medium", "low"]
